+++ daemon/daemon.go
@@ -1,0 +1,2 @@
+// Package daemon contains daemon types and functions
+package daemon