+++ io/write_test.go
@@ -1,0 +1,8 @@
+package io
+
+import "testing"
+
+func TestSaveJSON(t *testing.T) {
+	// TODO: SaveJSON
+	// TODO: SaveIndentedJSON
+}