+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## v0.3.9 - Daemon Package 
   * Commit: 
   * Update Root Package
+  * Dict SyncMap
+  * Daemon Instance
 ## v0.3.8 - Root Package
   * Commit: 2026-04-07 15:13
   * New Commands