+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   * Dict SyncMap
   * Daemon Instance
   * File System Functions
+  * Read and Write JSON
 ## v0.3.8 - Root Package
   * Commit: 2026-04-07 15:13
   * New Commands